Support synchronized method execution.
The `evaluate` struct and `Middle.cmd` channel allow functions of
type `func() error` to be channeled through the Middle's main logic
loop and safely access the Middle's data structures.  Currently
the only thing that implements this feature is the `RegisterConn`
method, which allows a user to initialie and add new connections,
but informs them when there's an error.