Significant refactor and design change.
The new wemdigo Middle design essentially consists of anonymous
connections, named target connections, and broadcast hubs.  Target
connections can be directly sent messages by the user via their string
identifier.  Each target has associated with it a hub, which is just a
set of connections.  Not every hub needs to be connected to a target,
and directing a message to a hub broadcasts the message to all of the
hub's members.